cleanup and readme file
